+++ tests/TestCase/Mailer/EmailTest.php
@@ -12,13 +12,13 @@
  * @since         2.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Test\TestCase\Network\Email;
+namespace Cake\Test\TestCase\Mailer;
 
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
 use Cake\Log\Log;
-use Cake\Network\Email\DebugTransport;
-use Cake\Network\Email\Email;
+use Cake\Mailer\Email;
+use Cake\Mailer\Transport\DebugTransport;
 use Cake\ORM\TableRegistry;
 use Cake\TestSuite\TestCase;
 use Exception;
@@ -660,7 +660,7 @@ class EmailTest extends TestCase
         $this->assertSame($expected, $this->CakeEmail->getHeaders(['from' => true, 'to' => true]));
 
         $result = $this->CakeEmail->setHeaders([]);
-        $this->assertInstanceOf('Cake\Network\Email\Email', $result);
+        $this->assertInstanceOf('Cake\Mailer\Email', $result);
     }
 
     /**
@@ -771,9 +771,9 @@ class EmailTest extends TestCase
         $this->assertSame($this->CakeEmail, $result);
 
         $result = $this->CakeEmail->transport();
-        $this->assertInstanceOf('Cake\Network\Email\DebugTransport', $result);
+        $this->assertInstanceOf('Cake\Mailer\Transport\DebugTransport', $result);
 
-        $instance = $this->getMock('Cake\Network\Email\DebugTransport');
+        $instance = $this->getMock('Cake\Mailer\Transport\DebugTransport');
         $this->CakeEmail->transport($instance);
         $this->assertSame($instance, $this->CakeEmail->transport());
     }
@@ -988,7 +988,7 @@ class EmailTest extends TestCase
         $this->assertEquals($config['theme'], $result);
 
         $result = $this->CakeEmail->transport();
-        $this->assertInstanceOf('Cake\Network\Email\DebugTransport', $result);
+        $this->assertInstanceOf('Cake\Mailer\Transport\DebugTransport', $result);
 
         $result = $this->CakeEmail->helpers();
         $this->assertEquals($config['helpers'], $result);
@@ -1646,7 +1646,7 @@ class EmailTest extends TestCase
         $this->CakeEmail->viewVars(['time' => $timestamp]);
 
         $result = $this->CakeEmail->helpers(['Time']);
-        $this->assertInstanceOf('Cake\Network\Email\Email', $result);
+        $this->assertInstanceOf('Cake\Mailer\Email', $result);
 
         $result = $this->CakeEmail->send();
         $dateTime = new \DateTime;
@@ -1823,7 +1823,7 @@ class EmailTest extends TestCase
     public function testDeliver()
     {
         $instance = Email::deliver('all@cakephp.org', 'About', 'Everything ok', ['from' => 'root@cakephp.org'], false);
-        $this->assertInstanceOf('Cake\Network\Email\Email', $instance);
+        $this->assertInstanceOf('Cake\Mailer\Email', $instance);
         $this->assertSame($instance->to(), ['all@cakephp.org' => 'all@cakephp.org']);
         $this->assertSame($instance->subject(), 'About');
         $this->assertSame($instance->from(), ['root@cakephp.org' => 'root@cakephp.org']);
@@ -2032,7 +2032,7 @@ class EmailTest extends TestCase
         $this->assertEquals($configs['subject'], $result);
 
         $result = $this->CakeEmail->transport();
-        $this->assertInstanceOf('Cake\Network\Email\DebugTransport', $result);
+        $this->assertInstanceOf('Cake\Mailer\Transport\DebugTransport', $result);
 
         $result = $this->CakeEmail->send('This is the message');
 
@@ -2088,7 +2088,7 @@ class EmailTest extends TestCase
         $this->assertEquals($configs['subject'], $result);
 
         $result = $this->CakeEmail->transport();
-        $this->assertInstanceOf('Cake\Network\Email\DebugTransport', $result);
+        $this->assertInstanceOf('Cake\Mailer\Transport\DebugTransport', $result);
 
         $result = $this->CakeEmail->send('This is the message');
 
@@ -2107,7 +2107,7 @@ class EmailTest extends TestCase
         $this->assertEquals('Cake\View\View', $result);
 
         $result = $this->CakeEmail->viewRender('Cake\View\ThemeView');
-        $this->assertInstanceOf('Cake\Network\Email\Email', $result);
+        $this->assertInstanceOf('Cake\Mailer\Email', $result);
 
         $result = $this->CakeEmail->viewRender();
         $this->assertEquals('Cake\View\ThemeView', $result);
@@ -2124,7 +2124,7 @@ class EmailTest extends TestCase
         $this->assertEquals('text', $result);
 
         $result = $this->CakeEmail->emailFormat('html');
-        $this->assertInstanceOf('Cake\Network\Email\Email', $result);
+        $this->assertInstanceOf('Cake\Mailer\Email', $result);
 
         $result = $this->CakeEmail->emailFormat();
         $this->assertEquals('html', $result);
@@ -2563,7 +2563,7 @@ HTML;
      */
     public function testMockTransport()
     {
-        $mock = $this->getMock('\Cake\Network\Email\AbstractTransport');
+        $mock = $this->getMock('\Cake\Mailer\AbstractTransport');
         $config = ['from' => 'tester@example.org', 'transport' => 'default'];
 
         Email::config('default', $config);
